Fix narrowing conversion

+++ src/main.cpp
@@ -27,7 +27,7 @@ int main()
     const int purple = 13;
 
     const unsigned int pageStartOffsets[2] = {
-        0, long(SCREEN_HEIGHT * SCREEN_WIDTH / 4)
+        0, SCREEN_HEIGHT * SCREEN_WIDTH / 4
     };
     
     int displayedPage = 0;
